+++ TVING/TVING/Present/Login/View/LoginView.swift
@@ -11,22 +11,24 @@ import SnapKit
 import Then
 
 final class LoginView: UIView {
-
+    
     // MARK: - Properties
     
     let loadUnderLine: [NSAttributedString.Key: Any] = [.underlineStyle: NSUnderlineStyle.single.rawValue]
     
     // MARK: - UI Components
     
-    private let preButton = UIButton()
     private let mainLabel = UILabel()
-    private let idTextField = UITextField()
-    private let passwordTextField = UITextField()
-    private let loginButton = UIButton()
+    let idTextField = UITextField()
+    let passwordTextField = UITextField()
+    let loginButton = UIButton()
     private let idFindButton = UIButton()
     private let passwordFindButton = UIButton()
     private let nonAccountLabel = UILabel()
-    private let nicknameButton = UIButton()
+    let nicknameButton = UIButton()
+    
+    var eyeButton = UIButton(type: .custom)
+    private let clearButton = UIButton()
     
     // MARK: - Life Cycles
     
@@ -36,6 +38,7 @@ final class LoginView: UIView {
         setStyle()
         setHierarchy()
         setLayout()
+        setPasswordShownButtonImage()
     }
     
     @available(*, unavailable)
@@ -47,12 +50,7 @@ final class LoginView: UIView {
 // MARK: - Extensions
 
 private extension LoginView {
-
     func setStyle() {
-        preButton.do {
-            $0.setImage(UIImage(resource: .btnBefore), for: .normal)
-        }
-        
         mainLabel.do {
             $0.text = StringLiterals.Login.mainTitle
             $0.font = .pretendardFont(weight: 500, size: 23)
@@ -62,7 +60,8 @@ private extension LoginView {
         
         idTextField.placeholder = StringLiterals.Login.idTextField
         passwordTextField.placeholder = StringLiterals.Login.passwordTextField
-    
+        passwordTextField.isSecureTextEntry = true
+        
         [idTextField, passwordTextField].forEach {
             $0.do {
                 $0.changePlaceholderColor(forPlaceHolder: $0.placeholder ?? "", forColor: .tvingGray2)
@@ -71,6 +70,7 @@ private extension LoginView {
                 $0.backgroundColor = .tvingGray4
                 $0.layer.cornerRadius = 3
                 $0.addPadding(left: 17)
+                $0.clearButtonMode = .whileEditing
             }
         }
         
@@ -99,6 +99,7 @@ private extension LoginView {
             $0.font = .pretendardFont(weight: 600, size: 14)
             $0.textAlignment = .center
         }
+        
         let attributeString = NSMutableAttributedString(string: StringLiterals.Login.nickname, attributes: loadUnderLine)
         nicknameButton.do {
             $0.setAttributedTitle(attributeString, for: .normal)
@@ -109,47 +110,36 @@ private extension LoginView {
     }
     
     func setHierarchy() {
-        addSubviews(preButton, mainLabel, idTextField)
-        addSubviews(passwordTextField, loginButton, idFindButton, passwordFindButton, nonAccountLabel, nicknameButton)
+        addSubviews(mainLabel, idTextField, passwordTextField, loginButton, idFindButton, passwordFindButton, nonAccountLabel, nicknameButton)
     }
     
     func setLayout() {
-        preButton.snp.makeConstraints {
-            $0.top.equalToSuperview().inset(65)
-            $0.leading.equalToSuperview().inset(24)
-            $0.height.equalTo(15)
-            $0.width.equalTo(8)
-        }
-        
         mainLabel.snp.makeConstraints {
-            $0.top.equalTo(preButton.snp.bottom).offset(10)
+            $0.top.equalToSuperview().inset(90)
             $0.leading.equalToSuperview().inset(109)
         }
         
         idTextField.snp.makeConstraints {
             $0.top.equalTo(mainLabel.snp.bottom).offset(31)
-            $0.left.equalToSuperview().inset(20)
-            $0.width.equalTo(335)
-            $0.height.equalTo(52)
         }
         passwordTextField.snp.makeConstraints {
             $0.top.equalTo(idTextField.snp.bottom).offset(7)
-            $0.left.equalToSuperview().inset(20)
-            $0.width.equalTo(335)
-            $0.height.equalTo(52)
         }
         
         loginButton.snp.makeConstraints {
             $0.top.equalTo(passwordTextField.snp.bottom).offset(21)
-            $0.left.equalToSuperview().inset(20)
-            $0.width.equalTo(335)
-            $0.height.equalTo(52)
         }
         
+        [idTextField, passwordTextField, loginButton].forEach {
+            $0.snp.makeConstraints {
+                $0.leading.trailing.equalToSuperview().inset(20)
+                $0.height.equalTo(52)
+            }
+        }
         
         idFindButton.snp.makeConstraints {
             $0.top.equalTo(loginButton.snp.bottom).offset(31)
-            $0.leading.equalToSuperview().inset(85)
+            $0.leading.equalToSuperview().inset(92)
             $0.height.equalTo(22)
         }
         
@@ -158,17 +148,35 @@ private extension LoginView {
             $0.leading.equalTo(idFindButton.snp.trailing).offset(69)
             $0.height.equalTo(22)
         }
-      
+        
         nonAccountLabel.snp.makeConstraints {
             $0.top.equalTo(idFindButton.snp.bottom).offset(28)
-            $0.leading.equalToSuperview().inset(51)
-           
+            $0.leading.equalToSuperview().inset(58)
         }
         
         nicknameButton.snp.makeConstraints {
-            $0.top.equalTo(nonAccountLabel.snp.top)
+            $0.top.equalTo(passwordFindButton.snp.bottom).offset(28)
             $0.leading.equalTo(nonAccountLabel.snp.trailing).offset(17)
             $0.height.equalTo(22)
+            $0.width.equalTo(128)
         }
+    }
+    
+    func setPasswordShownButtonImage() {
+        eyeButton = UIButton(primaryAction: UIAction( handler: { [self]_ in
+            passwordTextField.isSecureTextEntry.toggle()
+            self.eyeButton.isSelected.toggle()
+        }))
+        
+        var buttonConfig = UIButton.Configuration.plain()
+        buttonConfig.imagePadding = 10
+        buttonConfig.baseBackgroundColor = .clear
+        
+        eyeButton.setImage(UIImage(resource: .icSlasheye), for: .normal)
+        self.eyeButton.setImage(UIImage(resource: .icEye), for: .selected)
+        self.eyeButton.configuration = buttonConfig
+        
+        passwordTextField.rightView = eyeButton
+        passwordTextField.rightViewMode = .always
     }
 }
