+++ TVING/TVING/Present/Welcome/View/WelcomeView.swift
@@ -36,7 +36,6 @@ final class WelcomeView: UIView {
     required init?(coder: NSCoder) {
         fatalError("init(coder:) has not been implemented")
     }
-        
 }
 
 // MARK: - Extensions
